Fixing some issues on poisson-blending.py. More below.
Removing the landmark model options "mediapipe" and "spiga", since
they are not implemented yet. Also, zeroing the right image when
opening the UI to define the blending region.

Also, fixing the checkpoint and reconstruction step
configurations. The reconstruction steps were not used, since it was
tied to the checkpoint steps. Also, the actual model checkpoint was
not saved, so I fixed this.