Bugfix when reshaping the input coords to the image size. More below
The input coordinates for `ImageDataset.pixels` may not match the
original image size, for instance, in the Sandy case, the image size
is 655x655 pixels and we usually ask for 512x512 images on the
intermediate inferences, which led to exceptions being thrown. To fix
this, we need to simply reshape the integer coordinates using its own
dimensions. Problem is, we always assume that the image is square.